Add files via upload
An attention model is a method that takes n arguments y_1, ..., y_n (in the precedent examples, the y_i would be the h_i), and a context c. It return a vector z which is supposed to be the « summary » of the y_i, focusing on information linked to the context c. More formally, it returns a weighted arithmetic mean of the y_i, and the weights are chosen according the relevance of each y_i given the context c.
https://blog.heuritech.com/2016/01/20/attention-mechanism/
https://heuritech.files.wordpress.com/2016/01/detail_attentionmodel_step01.png?w=768